move the job label test into its own module
